src/sender/requirements.txt
mido==1.3.2
pygame==2.5.2
fluidsynth==0.2
pyfluidsynth==1.3.3
paho-mqtt==1.6.1
protobuf==3.12.4
ntplib
ntpserver